Lodha Codename Evergreen Vikhroli Mumbai
Relish luxury living amid green heaven of select peace and calm. Lodha Codename Evergreen– is a housing community that remains wanted in every way. Situated at Vikhroli this community has a place that has a superlative vantage point with connectivity from the best zones of Mumbai. 

Lodha Codename Vikhroli Mumbai project is spread over wide acres of land and represents a grand lifestyle, standing up strong and tall against the skyline. These luxury homes are Vaastu Compliant, smartly deliberate, and offer absolute beauty and grace.

As always, Lodha Group presents yet another novel theme-based development that truly précises the characteristic essence and special forecasts of home seekers who are dedicated to the world of luxury and relaxation as a delay of their lifestyle.

Your dream home not only with daylight & good exposure but also convenient with the best stipulations at the best price & supple payment plans. Clubhouse in Flats in Lodha Codename Evergreen Mumbai with a wide range of choices for people in your family is sure to be a hotspot. 

Rise exciting evenings or host big & small meetings at the party hall in the modern clubhouse. The kids can play in the kids' play area or ground or head to the kids’ pool for an exciting time.

The unparalleled location makes sure Lodha Flats in Vikhroli Mumbai place is the most preferred address for the next group. The protuberant connectivity and trouble-free access to other parts of the city include living here spacious.
The residential apartments are shaped with love and care to spring your all-out well-being in well-developed premises. These suave homes are no less than an absolution and illustrate the absolute fineness of elegance. They are enveloped by residential group covers and housing ventures.

The prices vary according to the formats, and you may explore the conveniences that it provides. This comforts the residents to get across to the significant location in and around the city seamlessly.

For More Details :
Visit - https://lodha.developerprojects.in/residential/mumbai/lodha-codename-evergreen-vikhroli/